requirements.txt
uvicorn==0.34.2
fastapi==0.115.12
pydantic==2.11.3
pydantic-settings==2.9.1
email_validator==2.2.0
sqlalchemy==2.0.40
passlib==1.7.4
psycopg2-binary==2.9.10
python_jose==3.4.0